[compiler] NFC: Move op equivalence into a file in Utils/
This is in preparation for reuse in CodeGen executable linking.
